Remove redundant type conversions

+++ pkg/util/util.go
@@ -109,7 +109,7 @@ func GetAvailableSpace(path string) (int64, error) {
 	if err != nil {
 		return int64(-1), err
 	}
-	return int64(stat.Bavail) * int64(stat.Bsize), nil
+	return int64(stat.Bavail) * stat.Bsize, nil
 }
 
 // GetAvailableSpaceBlock gets the amount of available space at the block device path specified.
